Itemsの削除APIで他のユーザーが作成したItemを削除できないようにする処理を追加

+++ src/features/items/items.service.ts
@@ -50,8 +50,11 @@ export class ItemsService {
     return await this.itemsRepository.save(item);
   }
 
-  async delete(id: string): Promise<void> {
+  async delete(id: string, user: User): Promise<void> {
     const item = await this.findById(id);
+    if (item.userId !== user.id) {
+      throw new BadRequestException('他人の商品を削除することはできません。');
+    }
     await this.itemsRepository.delete(item);
   }
 }
